discord-recovery-bot/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiofiles>=23.2.1
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
jinja2>=3.1.3
